setup: authentication middleware for tokens
