+++ middleware.ts
@@ -1,24 +1,24 @@
-import { NextRequest } from "next/server";
+import { NextRequest, NextResponse } from "next/server";
 
 export default function middleware(req: NextRequest) {
-    // const { nextUrl } = req;
-    // const token = req.cookies.get("next-auth.session-token") || req.cookies.get("__Secure-next-auth.session-token");
+    const { nextUrl } = req;
+    const token = req.cookies.get("next-auth.session-token") || req.cookies.get("__Secure-next-auth.session-token");
 
-    // const protectedRoutes = ["/dashboard", "/inventory"];
-    // const publicRoutes = ["/signin", "/signup", "/welcome"];
+    const protectedRoutes = ["/dashboard", "/inventory"];
+    const publicRoutes = ["/signin", "/signup", "/welcome"];
 
-    // const isProtectedRoute = protectedRoutes.some(route => nextUrl.pathname.startsWith(route));
-    // const isPublicRoute = publicRoutes.some(route => nextUrl.pathname.startsWith(route));
+    const isProtectedRoute = protectedRoutes.some(route => nextUrl.pathname.startsWith(route));
+    const isPublicRoute = publicRoutes.some(route => nextUrl.pathname.startsWith(route));
 
-    // if (isProtectedRoute && !token) {
-    //     return NextResponse.redirect(new URL("/signin", req.url));
-    // }
+    if (isProtectedRoute && !token) {
+        return NextResponse.redirect(new URL("/signin", req.url));
+    }
 
-    // if (isPublicRoute && token) {
-    //     return NextResponse.redirect(new URL("/dashboard", req.url));
-    // }
+    if (isPublicRoute && token) {
+        return NextResponse.redirect(new URL("/dashboard", req.url));
+    }
 
-    // return NextResponse.next();
+    return NextResponse.next();
 }
 
 export const config = {
