+++ package.json
@@ -14,6 +14,7 @@
     "bcryptjs": "^3.0.2",
     "chart.js": "^4.4.8",
     "cheerio": "^1.0.0",
+    "cloudinary": "^2.6.0",
     "clsx": "^2.1.1",
     "jsonwebtoken": "^9.0.2",
     "lucide-react": "^0.476.0",
