+++ server/controllers/dashboard.controller.ts
@@ -1,0 +1,9 @@
+import { NextRequest, NextResponse } from "next/server";
+
+export const editDashboardHandler = async (req: NextRequest) => {
+  console.log(req.user);
+  
+  return NextResponse.json({
+    h: "these are users",
+  });
+};
