+++ server/types/db/interfaces.ts
@@ -4,3 +4,7 @@ export interface CachedConnection {
     conn: Connection | null;
     promise: Promise<Connection> | null;
 }
+
+export interface PlatformUrlConfig {
+    [key: string]: (username: string) => string;
+}
